fix vec compare bugs in rv64

+++ rvp64_exec/builtin-rvp64-kdmtt16.c
@@ -6,7 +6,7 @@
 #include <rvp_intrinsic.h>
 #include <stdlib.h>
 #include <stdint.h>
-
+#include "test_utils.h"
 
 static __attribute__ ((noinline))
 uint64_t kdmtt16 (uint64_t ra, uint64_t rb)
@@ -23,7 +23,7 @@ int32x2_t v_kdmtt16 (int16x4_t ra, int16x4_t rb)
 int
 main ()
 {
-  int32_t a = kdmtt16 (0x0001f0000001f000, 0x0001100000011000);
+  uint64_t a = kdmtt16 (0x0001f0000001f000, 0x0001100000011000);
   int32x2_t va = v_kdmtt16 ((int16x4_t) {0xf777, 0xf111, 0xf777, 0xf111},
 			    (int16x4_t) {0x1000, 0x2000, 0x1000, 0x2000});
   if (a != 0x0000000200000002)
