+++ rvp64_exec/builtin-rvp64-kwmmul.c
@@ -6,10 +6,10 @@
 #include <rvp_intrinsic.h>
 #include <stdlib.h>
 #include <stdint.h>
-
+#include "test_utils.h"
 
 static __attribute__ ((noinline))
-int32_t kwmmul (int32_t ra, int32_t rb)
+int64_t kwmmul (int64_t ra, int64_t rb)
 {
   return __rv_kwmmul (ra, rb);
 }
@@ -17,7 +17,7 @@ int32_t kwmmul (int32_t ra, int32_t rb)
 int
 main ()
 {
-  int32_t a = kwmmul (0x8000000080000000, 0x8000000080000000);
+  int64_t a = kwmmul (0x8000000080000000, 0x8000000080000000);
 
   if (a != 0x7fffffff7fffffff)
     abort ();
